FIXES for compilation and execution on MACOSX
Some Makefile-fu to make compilation work, and get the right headers, use the right libraries, etc.

audio_portaudio.c was modified to use just one pablio "stream" for reading and writing (it keeps things separated on it's own), which also has been tested to work on Linux (I'll test Win32 soon).  I think this was how it was designed/intended to be used anyway.


git-svn-id: https://appconference.svn.sourceforge.net/svnroot/appconference/trunk@42 f854266a-e20d-0410-a9be-8852184c940f