Modularize the event callback mechanism to use either PostEvent, or the Immediate processing model with GuiLock that we'd been using.
The GuiLock stuff was causing the multiple register windows (on Win32
only), because the menu select event was getting processed over and
over, until one of the handlers finally returned, and took it off of the
event queue.

With PostEvent processing for the callback, this doesn't happen anymore,
because the queue isn't getting run from other threads.

Also, resolve (with a hack) the hang on exit experienced with windows.


git-svn-id: https://appconference.svn.sourceforge.net/svnroot/appconference/trunk@133 f854266a-e20d-0410-a9be-8852184c940f