Export local changes.
The biggest change here is that this now requires post-1.0 asterisk, and
uses the ast_frame->delivery mechanism to set timestamps.  Between that and
other changes in app_conference, this version is much more resilient to timing issues.

Any timing delays caused by scheduling get cleaned up nicely by remote jitterbuffers.

Also, we now implement a poor-mans "interpolation" to deal with lost frames [we generally use GSM only,
so this is just a repeat of a previous frame].


git-svn-id: https://appconference.svn.sourceforge.net/svnroot/appconference/trunk@272 f854266a-e20d-0410-a9be-8852184c940f