Fix a huge bug in the histogram jitterbuffer history mechanism, where we were looking past the end of the array of short-term history.  This make us get non-sense for "max".
With that fixed, I also changed it to try to tighten the length of the buffer, by using only one extra
frame of "fudge" (was 3), and allowing 2% drop from history.  (combined, this will usually result in
no loss, unless those 2% dropped are more than 1 frame late, and then we still have hysteresis in the
shrink/grow logic).


git-svn-id: https://appconference.svn.sourceforge.net/svnroot/appconference/trunk@343 f854266a-e20d-0410-a9be-8852184c940f