From: Cristian Draghici <cristian.draghici@gmail.com>
Hi Steve

Attached is a patch over Dominik's new registration code that adds the
message waiting indicator to the registration event structure.
I think it makes sense to add it in there as the MWI is sent within
the registration acknowledgement frame and it is dependent on the
registration id (i.e. each account has its own MWI value)

When the account used to register does not have an associated mailbox
the value of the msgcount is -1 (default value from lower level IES).

Cheers,
Cristi


git-svn-id: https://appconference.svn.sourceforge.net/svnroot/appconference/trunk@485 f854266a-e20d-0410-a9be-8852184c940f